requirements.txt
python_dotenv
openai
langchain
langchain_openai
langchain-community
youtube_transcript_api
tiktoken